Removed amp-state
Sending GET requests for amp-state requires a CORS URL, which Google sheet apps can't supply on the server-side, they only instruct to <a href="https://developers.google.com/api-client-library/javascript/features/cors" target="_blank">send an xhr request</a>. The current workaround is to create a "middleware" URL within the same domain that add the necessary CORS headers and any user authentication, but that can slow things down. The AMP developers are hesitant to fix it because AMP pages can be cached in multiple domains, which means the headers would have to relax to the point where it's not secure to work properly. Shelving this idea, but I've kept the hearts because I did get it to send a GA event.